+++ include/Dynamic.h
@@ -73,6 +73,13 @@ public:
 		return t==vtInt || t==vtFloat;
 	}
 
+	inline bool IsBool() const
+	{
+		if (!mPtr) return false;
+		int t = mPtr->__GetType();
+		return t==vtBool;
+	}
+
 
    int Compare(const Dynamic &inRHS) const
    {
@@ -87,10 +94,10 @@ public:
 
    bool operator != (const Dynamic &inRHS) const { return (Compare(inRHS) != 0); }
    bool operator != (const String &inRHS)  const { return !mPtr || ((String)(*this) != inRHS); }
-   bool operator != (double inRHS)  const { return !mPtr || ((double)(*this) != inRHS); }
-   bool operator != (float inRHS)  const { return !mPtr || ((double)(*this) != inRHS); }
-   bool operator != (int inRHS)  const { return !mPtr || ((double)(*this) != (double)inRHS); }
-   bool operator != (bool inRHS)  const { return !mPtr || ((double)(*this) != (double)inRHS); }
+   bool operator != (double inRHS)  const { return !IsNumeric() || ((double)(*this) != inRHS); }
+   bool operator != (float inRHS)  const { return !IsNumeric() || ((double)(*this) != inRHS); }
+   bool operator != (int inRHS)  const { return !IsNumeric() || ((double)(*this) != (double)inRHS); }
+   bool operator != (bool inRHS)  const { return !IsBool() || ((double)(*this) != (double)inRHS); }
 
    bool operator == (const Dynamic &inRHS) const
    {
@@ -101,10 +108,10 @@ public:
 
    #define DYNAMIC_COMPARE_OP( op ) \
       bool operator op (const String &inRHS)  const { return mPtr && ((String)(*this) op inRHS); } \
-      bool operator op (double inRHS)  const { return mPtr && ((double)(*this) op inRHS); } \
-      bool operator op (float inRHS)  const { return mPtr && ((double)(*this) op inRHS); } \
-      bool operator op (int inRHS)  const { return mPtr && ((double)(*this) op (double)inRHS); } \
-      bool operator op (bool inRHS)  const { return mPtr && ((double)(*this) op (double)inRHS); } \
+      bool operator op (double inRHS)  const { return IsNumeric() && ((double)(*this) op inRHS); } \
+      bool operator op (float inRHS)  const { return IsNumeric() && ((double)(*this) op inRHS); } \
+      bool operator op (int inRHS)  const { return IsNumeric() && ((double)(*this) op (double)inRHS); } \
+      bool operator op (bool inRHS)  const { return IsBool() && ((double)(*this) op (double)inRHS); } \
 
    #define DYNAMIC_COMPARE_OP_ALL( op ) \
       bool operator op (const Dynamic &inRHS) const { return mPtr && (Compare(inRHS) op 0); } \